Fix/mark height and weight as floating numbers (#48)
* Turn height and weight to numbers with floating point

* Fix: assign initial values in float

* Update UserProfileForm

* Fix tests

* Fix: bump simulator version

* Update Partial's c-tor: initialize fields safely through subscript method

* Fix: update ExampleApp to use the new constructor (+bugfix: timezone is not a string!)

* Update ios examples

* Fix: update tests

+++ ios/User/Sources/UserProfile.swift
@@ -12,8 +12,8 @@ public struct UserProfile: Codable {
     public let token: String
     public let birthDate: Date
     public let gender: Gender
-    public let height: Int
-    public let weight: Int
+    public let height: Float
+    public let weight: Float
     public var timezone: TimeZone { return TimeZone(identifier: _timezone)! }
     public let locale: String
